+++ src/http/handler/BaseHttpHandler.java
@@ -1,4 +1,4 @@
-package http;
+package http.handler;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
