+++ src/http/handler/PrioritizedHandler.java
@@ -1,4 +1,4 @@
-package http;
+package http.handler;
 
 import com.sun.net.httpserver.HttpExchange;
 import manager.TaskManager;
